+++ src/move-rows.js
@@ -1,65 +1,65 @@
 import { findIndex, omit, pick } from 'lodash';
-import { moveRows } from 'reactabular-dnd';
 
 const moveTreeRows = ({
-  sourceRowId,
-  targetRowId,
+  operation,
   retain = [],
   idField = 'id',
   parentField = 'parent'
-} = {}) => (rows) => {
-  let movedRows = moveRows({
-    sourceRowId,
-    targetRowId,
-    idField
-  })(rows);
-  let cancelMoving = false;
+} = {}) => {
+  if (!operation) {
+    throw new Error('tree.moveRows - Missing operation!');
+  }
 
-  if (movedRows) {
-    // Walk through the old row definition, patch parent relations and fields
-    // of the new one
-    movedRows = rows.map((row, i) => {
-      if (typeof row[parentField] === 'undefined' || row[parentField] === null) {
+  return (rows) => {
+    let movedRows = operation(rows);
+    let cancelMoving = false;
+
+    if (movedRows) {
+      // Walk through the old row definition, patch parent relations and fields
+      // of the new one
+      movedRows = rows.map((row, i) => {
+        if (typeof row[parentField] === 'undefined' || row[parentField] === null) {
+          return {
+            ...omit(movedRows[i], retain),
+            ...pick(row, retain),
+            [idField]: movedRows[i][idField],
+            [parentField]: undefined
+          };
+        }
+
+        // Find the index of the old parent
+        const index = findIndex(rows, {
+          [idField]: row[parentField]
+        });
+
+        if (index < 0) {
+          console.warn( // eslint-disable-line no-console
+            'tree.moveRows - Failed to find the old parent', rows, row, idField, parentField
+          );
+
+          cancelMoving = true;
+
+          return null;
+        }
+
+        // Figure out the new id based on that index
+        const id = movedRows[index][idField];
+
         return {
           ...omit(movedRows[i], retain),
           ...pick(row, retain),
           [idField]: movedRows[i][idField],
-          [parentField]: undefined
+          [parentField]: id
         };
-      }
-
-      // Find the index of the old parent
-      const index = findIndex(rows, {
-        [idField]: row[parentField]
       });
+    }
 
-      if (index < 0) {
-        console.warn( // eslint-disable-line no-console
-          'tree.moveRows - Failed to find the old parent', rows, row, idField, parentField
-        );
+    if (cancelMoving) {
+      return rows;
+    }
 
-        cancelMoving = true;
-
-        return null;
-      }
-
-      // Figure out the new id based on that index
-      const id = movedRows[index][idField];
-
-      return {
-        ...omit(movedRows[i], retain),
-        ...pick(row, retain),
-        [idField]: movedRows[i][idField],
-        [parentField]: id
-      };
-    });
-  }
-
-  if (cancelMoving) {
-    return rows;
-  }
-
-  return movedRows;
+    return movedRows;
+  };
 };
 
 export default moveTreeRows;